refactor: flatten material module
