+++ src/app/components/hours-dialog/hours-dialog.component.spec.ts
@@ -5,14 +5,12 @@ import { NoopAnimationsModule } from '@angular/platform-browser/animations';
 import { OverlayContainer } from '@angular/cdk/overlay';
 import { By } from '@angular/platform-browser';
 
-import { MaterialModule } from '@shared/material/material.module';
+import { SharedModule } from '@shared/shared-module.module';
 
 import { HoursDialogComponent } from './hours-dialog.component';
 
 // Noop component is only a workaround to trigger change detection
-@Component({
-  template: ''
-})
+@Component({})
 class NoopComponent { }
 
 const TEST_DIRECTIVES = [
@@ -52,4 +50,3 @@ describe('HoursDialogComponent', () => {
     dialog = TestBed.get(MatDialog);
   });
 });
-
