+++ src/app/modules/shared/components/content-with-side-nav/content-with-side-nav.component.spec.ts
@@ -1,7 +1,7 @@
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 
 import { ContentwithSideNavComponent } from './content-with-side-nav.component';
-import { MaterialModule } from '../../material/material.module';
+import { MaterialModule } from '../../material.module';
 import { RouterTestingModule } from '@angular/router/testing';
 import { NoopAnimationsModule } from '@angular/platform-browser/animations';
 
@@ -13,9 +13,9 @@ describe('ContentwithSideNavComponent', () => {
   beforeEach(async(() => {
     TestBed.configureTestingModule({
       imports: [MaterialModule, RouterTestingModule, NoopAnimationsModule],
-      declarations: [ ContentwithSideNavComponent ]
+      declarations: [ContentwithSideNavComponent]
     })
-    .compileComponents();
+      .compileComponents();
   }));
 
   beforeEach(() => {
